requirements.txt
numpy>=1.26
matplotlib>=3.10
scikit-learn>=1.3
joblib>=1.3
Pillow>=10.0
seaborn>=0.12